updateMessage
