+++ apps/meteor/app/apps/server/bridges/livechat.ts
@@ -74,7 +74,8 @@ export class AppLivechatBridge extends LivechatBridge {
 			message: await this.orch.getConverters()?.get('messages').convertAppMessage(message),
 		};
 
-		await Livechat.updateMessage(data);
+		// @ts-expect-error IVisitor vs ILivechatVisitor :(
+		await LivechatTyped.updateMessage(data);
 	}
 
 	protected async createRoom(visitor: IVisitor, agent: IUser, appId: string, extraParams?: IExtraRoomParams): Promise<ILivechatRoom> {
